🌱 Split cache.deepcover.test.ts into 3 focused test files
Split 1835-line cache.deepcover.test.ts into:
- cache.deepcover.core.test.ts (~401 lines): useCache hook, demo mode, enabled flag, failure tracking, refetch
- cache.deepcover.store.test.ts (~492 lines): CacheStore tests (merge, concurrent, reset, loadFromStorage, saveMeta)
- cache.deepcover.migration.test.ts (~1130 lines): localStorage/IDB migrations, preload, edge cases

All 89 tests pass. Part of #10439.
